feat: Specify dtype=float32 for translation column in user and article translations

+++ src/make_article_translation.py
@@ -14,7 +14,7 @@ def load_article_embeddings(file_path):
 def save_article_translation(df, translations, output_file):
     print("DataFrame shape before adding translations:", df.shape)
     print("Translations shape:", translations.shape)
-    df["translation"] = df.index.map(lambda idx: translations[idx])
+    df["translation"] = df.index.map(lambda idx: translations[idx]).astype(np.float32)
     df.to_parquet(output_file, index=False)
 
 def main(article_embeddings_file, article_translation_file, model_path):
